Change Integer.parseInt to Float.parseFloat

+++ hw6.java
@@ -1,7 +1,7 @@
 class hw6{
 	
 	 public static void main (String [] args) {
-		float TF = Integer.parseInt(args[0]);
+		float TF = Float.parseFloat(args[0]);
 		float TC;
 		TC = (TF - 32) * 5/9;
 		System.out.println ("The temperature is  " + TC + " degrees Celsius");
