+++ hw7.java
@@ -1,7 +1,7 @@
 class hw7 {
 	
 	public static void main (String [] args){
-		float TC = Integer.parseInt (args[0]);
+		float TC = Float.parseFloat (args[0]);
 		float TF;
 		TF = TC /5*9 +32;
 		System.out.println (" The temperature is " + TF + " degree Fahrenheit");
